requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
requests==2.28.2
urllib3==1.26.14
Werkzeug==2.2.3
zipp==3.14.0
python-dotenv==1.0.0
